yui-telegram-bot/requirements.txt
python-telegram-bot==13.15
openai==1.13.3
python-dotenv==1.0.1
